smp: fixes

+++ lib/http2/http2.c
@@ -1495,7 +1495,7 @@ lws_h2_parser(struct lws *wsi, unsigned char *in, lws_filepos_t inlen,
 				 * the frame boundary, in the case there is already
 				 * more waiting leave it for next time around
 				 */
-				n = inlen + 1;
+				n = (int)inlen + 1;
 				if (n > (int)(h2n->length - h2n->count + 1)) {
 					n = h2n->length - h2n->count + 1;
 					lwsl_debug("---- restricting len to %d vs %ld\n", n, (long)inlen + 1);
